requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
fsspec==2024.10.0
grpcio==1.67.1
grpcio-reflection==1.67.1
grpcio-tools==1.67.1
idna==3.10
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
protobuf==5.28.3
regex==2024.9.11
requests==2.32.3
setuptools==75.3.0
six==1.16.0
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3

soundfile~=0.12.1
vosk~=0.3.44